--- NewEventForm.js
@@ -1,3 +1,0 @@
-export default function NewEventForm(){
-return( )
-}
