+++ src/Components/Attendees.js
@@ -1,3 +1,35 @@
-export default function Attendees() {
-  return;
+
+import Attendee from "./Attendee";
+import { useState } from "react";
+
+export default function Attendees({event, attendees, updateEventAttendance}){
+
+  const [showAttendees, setShowAttendees] = useState(false)
+
+  function toggleEventAttendees (){
+    setShowAttendees(!showAttendees)
+  }
+    
+    return (
+<>
+                      <button onClick={toggleEventAttendees}>
+                        {!showAttendees ? "Show Attendees" : "Hide Attendees"}
+                      </button>
+
+                      {showAttendees ? (
+                        <div className="attendees">
+                          {attendees.map((attendee, index) => (
+                            
+                              <Attendee 
+                              attendee ={attendee}
+                              event ={event}
+                              updateEventAttendance ={updateEventAttendance}
+                              />
+                      
+                          ))}
+                        </div>
+                      ) : null}
+                    </>
+
+    )
 }
