--- src/Footer.js
@@ -1,11 +1,0 @@
-export default function Footer (){
-    return (<>
-      <footer>
-        <ul>
-          <li>Contact</li>
-          <li>About</li>
-          <li>Legal</li>
-        </ul>
-      </footer>
-    </>)
-}
