--- src/Header.js
@@ -1,7 +1,0 @@
-export default function Header(){
-    return (<>
-        <header>
-          <h1 className="color-change-5x">RSVP App</h1>
-        </header>
-      </>)
-}
